Added Global tracklist option
Added a flag in PPlusNetplayMusicPlayer.json to have a flag to set global music off or on. Music player py file now checks for this flag. If true, it checks if the global.tlst file exists and reads the global.tlst for songs on every stage rather than individual tlsts per stage. Menu and Results screen still separate.